+++ Ejercicio3/ApartadoC/src/IPromotionStrategy.java
@@ -1,5 +1,5 @@
 package Ejercicio3.ApartadoC;
 
 public interface IPromotionStrategy {
-    int applyPromotion(int basePrice);
+    double applyPromotion(double basePrice);
 }
